Split execution recovery modules
