+++ crates/oasis7/src/bin/oasis7_chain_runtime/execution_bridge/tests/mod.rs
@@ -24,6 +24,7 @@ use sha2::{Digest, Sha256};
 
 mod determinism;
 mod driver;
+mod driver_authoritative_recovery;
 mod driver_checkpoint;
 mod driver_checkpoint_profile;
 mod driver_observability;
